+++ include/audio/zoe_audio_io_proc_scene_menu_main.h
@@ -1,0 +1,10 @@
+#ifndef __zoe_audio_zoe_io_proc_scene_menu_main_h
+#define __zoe_audio_zoe_io_proc_scene_menu_main_h
+
+#include <metil_audio/metil_audio_io_proc.h>
+
+metil_audio_io_proc_macro_type(
+  zoe_audio_io_proc_scene_menu_main
+);
+
+#endif
