+++ include/scenes/scene_menu_main.h
@@ -28,17 +28,6 @@ enum textures_scene_menu_main {
   textures_scene_menu_main_menu_exit = 4
 };
 
-struct scene_menu_main_data {
-  struct metil_menu menu;
-  unsigned long int time_started;
-
-  struct rand_parameters rand_parameters;
-  struct rand_source rand_source;
-  struct rand_result rand_result;
-
-  struct io_proc_data* _Nonnull io_proc_data;
-};
-
 void scene_menu_main_initialize(
   struct metil* _Nonnull,
   struct metil_scene* _Nonnull
@@ -58,25 +47,5 @@ void scene_menu_main_destroy(
   struct metil* _Nonnull,
   struct metil_scene* _Nonnull
 );
-
-#if target_os_ios
-int scene_menu_main_io_proc(
-  unsigned char,
-  const AudioTimeStamp* _Nonnull,
-  unsigned int,
-  AudioBufferList* _Nonnull,
-  void* _Nonnull
-);
-#else
-OSStatus scene_menu_main_io_proc(
-  AudioObjectID,
-  const AudioTimeStamp* _Nonnull,
-  const AudioBufferList* _Nonnull,
-  const AudioTimeStamp* _Nonnull,
-  AudioBufferList* _Nonnull,
-  const AudioTimeStamp* _Nonnull,
-  void* _Nonnull
-);
-#endif
 
 #endif
