+++ sources/scenes/scene_menu_main.m
@@ -1,6 +1,8 @@
 #include <scenes/scene_menu_main.h>
 
+#include <audio/zoe_audio_io_proc_scene_menu_main.h>
 #include <data/data_zoe.h>
+#include <data/zoe_data_scene_menu_main.h>
 #include <menus/menu_main.h>
 #include <object/object_ground.h>
 #include <object/object_tree.h>
@@ -38,15 +40,6 @@
 #include <rand_result.h>
 #include <rand_source.h>
 #include <rand_source_type.h>
-
-#if target_os_ios
-#include <AVFAudio/AVFAudio.h>
-#else
-#include <CoreAudio/CoreAudio.h>
-#endif
-const unsigned long int scene_menu_main_time_scene_transition = (
-  3333
-);
 
 void scene_menu_main_initialize(
   struct metil* metil,
@@ -89,7 +82,7 @@ void scene_menu_main_initialize(
 
   metil_audio_io_proc_add_with_data(
     &metil->audio,
-    scene_menu_main_io_proc,
+    zoe_audio_io_proc_scene_menu_main,
     io_proc_data
   );
 
@@ -462,282 +455,3 @@ void scene_menu_main_destroy(
     scene
   );
 }
-
-#if target_os_ios
-int scene_menu_main_io_proc(
-  unsigned char silence,
-  const AudioTimeStamp* _Nonnull timestamp,
-  AVAudioFrameCount frame_count,
-  AudioBufferList* _Nonnull output_data,
-  void* data
-) {
-  struct metil_audio_io_proc_data* metil_audio_io_proc_data = (
-    data
-  );
-
-  struct io_proc_data* io_proc_data = (
-    metil_audio_io_proc_data->data
-  );
-
-  if (
-    io_proc_data->destroy == 1
-  ) {
-    metil_audio_io_proc_remove(
-      &metil_audio_io_proc_data->metil->audio,
-      scene_menu_main_io_proc
-    );
-
-    rand_clean(
-      &io_proc_data->rand_result,
-      &io_proc_data->rand_source
-    );
-
-    free(
-      io_proc_data
-    );
-
-    return 0;
-  }
-
-  struct metil_scene_controller* metil_scene_controller = (
-    metil_audio_io_proc_data->metil->scene_controller
-  );
-
-  struct metil_scene* metil_scene = &(
-    metil_scene_controller->scene
-  );
-
-  struct scene_menu_main_data* scene_menu_main_data = (
-    metil_scene->data
-  );
-
-  float volume_multiplier = (
-    1.0f
-  );
-
-  if (
-    scene_menu_main_data->time_started != 0
-  ) {
-    unsigned long int time_delta = (
-      metil_scene->time -
-      scene_menu_main_data->time_started
-    );
-
-    if (
-      time_delta >= scene_menu_main_time_scene_transition
-    ) {
-      volume_multiplier = (
-        0.0f
-      );
-    } else {
-      volume_multiplier = (
-        (
-          (float) (scene_menu_main_time_scene_transition - time_delta) /
-          (float) scene_menu_main_time_scene_transition
-        ) *
-        0.75 +
-        0.25f
-      );
-    }
-  }
-
-  rand_get(
-    &io_proc_data->rand_source,
-    &io_proc_data->rand_result,
-    &io_proc_data->rand_parameters
-  );
-
-  for (
-    unsigned long int index_buffer = 0;
-    index_buffer < output_data->mNumberBuffers;
-    ++index_buffer
-  ) {
-    AudioBuffer audio_buffer_current = output_data->mBuffers[
-      index_buffer
-    ];
-
-    float* buffer_out = audio_buffer_current.mData;
-    unsigned long int count_channel_out = audio_buffer_current.mNumberChannels;
-
-    for (
-      unsigned int index_frame = 0;
-      index_frame < frame_count;
-      ++index_frame
-    ) {
-      unsigned long int channel = (
-        index_frame %
-        count_channel_out
-      );
-
-      unsigned int offset_byte = (
-        index_frame *
-        2
-      );
-
-      buffer_out[
-        index_frame
-      ] = (
-        (
-          (float) (
-            (
-              io_proc_data->rand_result.bytes[
-                offset_byte %
-                20500
-              ] *
-              io_proc_data->rand_result.bytes[
-                (
-                  offset_byte +
-                  1
-                ) %
-                20500
-              ]
-            ) %
-            10000
-          )
-        ) /
-        100000.0f *
-        volume_multiplier
-      );
-    }
-  }
-
-  return 0;
-}
-#else
-OSStatus scene_menu_main_io_proc(
-  AudioObjectID id_audio_object,
-  const AudioTimeStamp* time_stamp_audio,
-  const AudioBufferList* list_buffer_audio_in,
-  const AudioTimeStamp* time_stamp_audio_in,
-  AudioBufferList* list_buffer_audio_out,
-  const AudioTimeStamp* time_stamp_audio_out,
-  void* data
-) {
-  struct metil_audio_io_proc_data* metil_audio_io_proc_data = (
-    data
-  );
-
-  struct io_proc_data* io_proc_data = (
-    metil_audio_io_proc_data->data
-  );
-
-  if (
-    io_proc_data->destroy == 1
-  ) {
-    metil_audio_io_proc_remove(
-      &metil_audio_io_proc_data->metil->audio,
-      scene_menu_main_io_proc
-    );
-
-    rand_clean(
-      &io_proc_data->rand_result,
-      &io_proc_data->rand_source
-    );
-
-    free(
-      io_proc_data
-    );
-
-    return 0;
-  }
-
-  struct metil_scene_controller* metil_scene_controller = (
-    metil_audio_io_proc_data->metil->scene_controller
-  );
-
-  struct metil_scene* metil_scene = &(
-    metil_scene_controller->scene
-  );
-
-  struct scene_menu_main_data* scene_menu_main_data = (
-    metil_scene->data
-  );
-
-  float volume_multiplier = (
-    1.0f
-  );
-
-  if (
-    scene_menu_main_data->time_started != 0
-  ) {
-    unsigned long int time_delta = (
-      metil_scene->time -
-      scene_menu_main_data->time_started
-    );
-
-    if (
-      time_delta >= scene_menu_main_time_scene_transition
-    ) {
-      volume_multiplier = (
-        0.0f
-      );
-    } else {
-      volume_multiplier = (
-        (
-          (float) (scene_menu_main_time_scene_transition - time_delta) /
-          (float) scene_menu_main_time_scene_transition
-        ) *
-        0.75 +
-        0.25f
-      );
-    }
-  }
-
-  rand_get(
-    &io_proc_data->rand_source,
-    &io_proc_data->rand_result,
-    &io_proc_data->rand_parameters
-  );
-
-  for (
-    unsigned long int index_buffer = 0;
-    index_buffer < list_buffer_audio_out->mNumberBuffers;
-    ++index_buffer
-  ) {
-    AudioBuffer audio_buffer_current = list_buffer_audio_out->mBuffers[index_buffer];
-
-    float* buffer_out = audio_buffer_current.mData;
-    unsigned long int size_buffer_out = audio_buffer_current.mDataByteSize / sizeof(float);
-    unsigned long int count_channel_out = audio_buffer_current.mNumberChannels;
-
-    for (
-      unsigned long int index_buffer_out = 0;
-      index_buffer_out < size_buffer_out;
-      ++index_buffer_out
-    ) {
-      unsigned long int channel = index_buffer_out % count_channel_out;
-      unsigned int offset_byte = (
-        index_buffer_out *
-        2
-      );
-
-      buffer_out[
-        index_buffer_out
-      ] = (
-        (
-          (float) (
-            (
-              io_proc_data->rand_result.bytes[
-                offset_byte %
-                20500
-              ] *
-              io_proc_data->rand_result.bytes[
-                (
-                  offset_byte +
-                  1
-                ) %
-                20500
-              ]
-            ) %
-            10000
-          )
-        ) /
-        100000.0f *
-        volume_multiplier
-      );
-    }
-  }
-
-  return 0;
-}
-#endif
